AutoHotkey Kee4 implementation
The language mode has been removed and the language key is used instead to
lock the Mode Select mode.

Three keys down aborts the current sequence and shows the HUD for the
current mode. Four keys down resets the current mode to Lowcase and shows
the HUD for all the other modes.

To repeat an action tap the last key that was released less than 200 ms
after it. This is how modes are locked to. The key can then be repeatedly
tapped or held down for auto-repeat.

The numpad can either be used in portrait or ladscape (Horizontal).