Bracket autocomplete and Mouse mode
Double tapping an opening bracket applies the coresponding closing
bracket. Similarly, a double tapped Space will replace the first space
with a fullstop and latch the Capitals mode.

To enter Mouse mode double tap the Nav key. Now that there are so many
modes, the current mode is omitted from the all-keys-down HUD.